Reorganize reducers
