+++ src/redux/reducers/ordered/sessionActivities.js
@@ -1,0 +1,29 @@
+import _ from 'lodash';
+import types from '../../actions/types';
+
+
+const sessionActivities = function(sessionActivities = {}, action) {
+  switch(action.type) {
+    case types.DELETE_SESSION: {
+
+    }
+
+    case types.ADD_SESSION: {
+
+    }
+
+    case types.ADD_ACTIVITY: {
+
+    }
+
+    case types.DELETE_SESSION: {
+
+    }
+
+    default:
+      return sessionActivities;
+  }
+};
+
+
+export default sessionActivities;
